+++ src/components/ui/Form/index.tsx
@@ -4,7 +4,6 @@ import { Slot } from '@radix-ui/react-slot';
 import {
   Controller,
   FormProvider,
-  useFormContext,
   type ControllerProps,
   type FieldPath,
   type FieldValues,
@@ -12,22 +11,13 @@ import {
 import { Label } from '@/components/ui/Label';
 import { formItem, formDescription, formMessage, errorLabel } from './form.css';
 import { cx } from '@/utils/cx';
-
+import { useFormField } from './hook';
+import { FormFieldContext, FormItemContext } from './context';
 /**
  * shadcn/form
  * @see https://ui.shadcn.com/docs/components/form
  */
-
 const Form = FormProvider;
-
-type FormFieldContextValue<
-  TFieldValues extends FieldValues = FieldValues,
-  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
-> = {
-  name: TName;
-};
-
-const FormFieldContext = React.createContext<FormFieldContextValue>({} as FormFieldContextValue);
 
 const FormField = <
   TFieldValues extends FieldValues = FieldValues,
@@ -41,35 +31,6 @@ const FormField = <
     </FormFieldContext.Provider>
   );
 };
-
-const useFormField = () => {
-  const fieldContext = React.useContext(FormFieldContext);
-  const itemContext = React.useContext(FormItemContext);
-  const { getFieldState, formState } = useFormContext();
-
-  const fieldState = getFieldState(fieldContext.name, formState);
-
-  if (!fieldContext) {
-    throw new Error('useFormField should be used within <FormField>');
-  }
-
-  const { id } = itemContext;
-
-  return {
-    id,
-    name: fieldContext.name,
-    formItemId: `${id}-form-item`,
-    formDescriptionId: `${id}-form-item-description`,
-    formMessageId: `${id}-form-item-message`,
-    ...fieldState,
-  };
-};
-
-type FormItemContextValue = {
-  id: string;
-};
-
-const FormItemContext = React.createContext<FormItemContextValue>({} as FormItemContextValue);
 
 const FormItem = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
   ({ className, ...props }, ref) => {
@@ -102,15 +63,6 @@ const FormLabel = React.forwardRef<
 });
 FormLabel.displayName = 'FormLabel';
 
-/**
- * FormControl은 input, select, textarea 등 다양한 입력 요소를 자식으로 받을 수 있어야 합니다.
- * Slot을 사용하면 어떤 컴포넌트든 자식으로 받아 속성을 전달할 수 있습니다.
- * 접근성 속성 주입, ID 연결이 가능합니다.
- * @example
- * <FormControl>
-      <Input {...field} /> {/* Input 컴포넌트가 Slot의 자식으로 들어갑니다
-   </FormControl>
- */
 const FormControl = React.forwardRef<HTMLSlotElement, React.ComponentPropsWithoutRef<typeof Slot>>(
   ({ ...props }, ref) => {
     const { error, formItemId, formDescriptionId, formMessageId } = useFormField();
@@ -159,13 +111,4 @@ const FormMessage = React.forwardRef<
 });
 FormMessage.displayName = 'FormMessage';
 
-export {
-  useFormField,
-  Form,
-  FormItem,
-  FormLabel,
-  FormControl,
-  FormDescription,
-  FormMessage,
-  FormField,
-};
+export { Form, FormItem, FormLabel, FormControl, FormDescription, FormMessage, FormField };
